Cargando avances viernes 0608

+++ database/migrations/2021_07_30_214148_create_avances_table.php
@@ -16,10 +16,10 @@ class CreateAvancesTable extends Migration
         Schema::create('avances', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->float('avance_plan');
-            $table->float('avance_real');
-            $table->float('avance_desviacion');
-            $table->float('avance_cumplimiento');
+            $table->integer('avance_plan');
+            $table->integer('avance_real');
+            $table->integer('avance_desviacion');
+            $table->integer('avance_cumplimiento');
             $table->unsignedBigInteger('asignacion_id');
             $table->foreign('asignacion_id')->references('id')->on('asignacions');
         });
